last vid update

+++ app/actions/general.ts
@@ -9,33 +9,28 @@ import { clerkClient } from "@clerk/nextjs/server";
 
 export async function deleteDataById(
   id: string,
+
   deleteType: "doctor" | "staff" | "patient" | "payment" | "bill"
 ) {
   try {
-    const client = await clerkClient();
-
     switch (deleteType) {
       case "doctor":
-        await db.doctor.delete({ where: { id } });
-        await client.users.deleteUser(id);
-        break;
-
+        await db.doctor.delete({ where: { id: id } });
       case "staff":
-        await db.staff.delete({ where: { id } });
-        await client.users.deleteUser(id);
-        break;
-
+        await db.staff.delete({ where: { id: id } });
       case "patient":
-        await db.patient.delete({ where: { id } });
-        await client.users.deleteUser(id);
-        break;
-
+        await db.patient.delete({ where: { id: id } });
       case "payment":
         await db.payment.delete({ where: { id: Number(id) } });
-        break;
+    }
 
-      default:
-        return { success: false, message: "Invalid delete type", status: 400 };
+    if (
+      deleteType === "staff" ||
+      deleteType === "patient" ||
+      deleteType === "doctor"
+    ) {
+      const client = await clerkClient();
+      await client.users.deleteUser(id);
     }
 
     return {
@@ -43,8 +38,9 @@ export async function deleteDataById(
       message: "Data deleted successfully",
       status: 200,
     };
-  } catch (error: any) {
-    console.log("DELETE ERROR:", error.message);
+  } catch (error) {
+    console.log(error);
+
     return {
       success: false,
       message: "Internal Server Error",
